+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.0.2
 
+- Move to using [nuxt-directus package](https://nuxt-directus.netlify.app/)
 - Add .DS_Store to gitignore
 
 ## 0.0.1